feat: migrated handler logic to separate folder
