+++ server/app.js
@@ -5,10 +5,10 @@ const asyncHandler = require('express-async-handler');
 const { check, validationResult } = require('express-validator');
 const { body } = require('express-validator');
 const HttpError = require("./utils/HttpError");
-const regionRouter = require("./routes/regionRouter"); // create your router
-const {errorHandler} = require("./routes/utils");
+const router = require("./routes"); // create your router
+const {errorHandler} = require("./handlers/utils");
 const log = require("loglevel");
-const helper = require("./routes/utils");
+const helper = require('./handlers/utils');
 const config = require('../config/config')
 
 const app = express();
@@ -31,7 +31,7 @@ app.use(bodyParser.urlencoded({ extended: false })); // parse application/x-www-
 app.use(bodyParser.json()); // parse application/json
 
 //routers
-app.use('/region', regionRouter);
+app.use('/', router);
 
 //paths
 //app.get('/entity', asyncHandler(async (req, res, next) => {
